requirements.txt
PyQt5
whisper
python-docx
fpdf
transformers
exceptions
dotenv